Refactoring, fix automations.

+++ helper_scripts/run_trials.sh
@@ -1,11 +1,18 @@
+#!/bin/bash
+
+if [ "$#" -ne 1 ]; then
+    echo "Usage: $0 <NUM OF NODES>"
+    exit 1
+fi
+
+N=$1
+
 for i in $(seq 1 20);
 do
-	pushd "$HOME"/fed_5g/helper_scripts || exit
-	./dataset_distributor.sh 3  # N - number of clients
-	cp cifar10_3_partitions_iid.png cifar10_3_partitions_iid_trial_$i.png
-	popd || exit
+	./dataset_distributor.sh ${N}
+	cp cifar10_"${N}"_partitions_iid.png cifar10_"${N}"_partitions_iid_trial_$i.png
 	sleep 2
-	flwr run . --stream
+	flwr run "$HOME"/fed_5g/flwr-application --stream
 	sleep 2
 	kill -9 $(pgrep tshark)
 done
